Перенесён класс `TemplateConsumable`, удалён излишний метод `GetElementIdValue`, рефакторинг обработки Id.
